requirements.txt
alembic==1.1.0
aniso8601==8.0.0
atomicwrites==1.3.0
attrs==19.1.0
backports.functools-lru-cache==1.5
certifi==2019.9.11
chardet==3.0.4
Click==7.0
contextlib2==0.5.5
docopt==0.6.2
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Login==0.4.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
future==0.17.1
idna==2.8
importlib-metadata==0.22
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==3.0.5
marshmallow-sqlalchemy==0.19.0
more-itertools==7.2.0
paho-mqtt==1.3.1
pbr==5.4.3
pluggy==0.13.0
py==1.8.0
PyJWT==1.7.1
pykwalify==1.7.0
pytest==4.5.0
python-box==3.4.3
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
PyYAML==5.1.2
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.8
stevedore==1.31.0
tavern==0.30.3
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.6
zipp==0.6.0
